+++ src/main/java/xyz/nikitacartes/easyauth/config/LangConfigV1.java
@@ -329,6 +329,8 @@ public class LangConfigV1 extends ConfigTemplate {
         public TranslatableText forceLoginSuccess = new TranslatableText("admin.forceLoginSuccess");
         public TranslatableText ipLimitNotify = new TranslatableText("admin.ipLimitNotify");
         public TranslatableText databaseUnavailable = new TranslatableText("admin.databaseUnavailable");
+        public TranslatableText accounts = new TranslatableText("admin.accounts");
+        public TranslatableText accountsNoIp = new TranslatableText("admin.accountsNoIp");
         public TranslatableText backupSuccess = new TranslatableText("admin.backupSuccess");
         public TranslatableText backupUnsupported = new TranslatableText("admin.backupUnsupported");
     }